requirements.txt
# Optional: JSON Patch (RFC 6902) for `patches/agenticseek` manifest `jsonpatch` steps.
jsonpatch>=1.33

# Optional runtime deps for `python3 -m local_ai_stack status` (Rich dashboard).
rich>=13.7.0

# Interactive remediation Doctor UI (`python -m claude_bridge.remediation_tui` / `local_ai_stack remediation-ui`).
textual>=0.47.0

# Fopoon dashboard (`python -m ui.dashboard`): host metrics + Docker + trace tail.
psutil>=6.0.0
nvidia-ml-py>=12.0.0

# MCP server for Turntable / CTI Pilot tools (`PYTHONPATH=src python -m octo_spork_mcp.octo_tools`).
mcp>=1.2.0

# GitHub PR comment bot (`python github_integrator.py`).
# Optional system package: Graphviz (`dot` on PATH) for `repo_graph.dot_svg` dependency diagrams.
PyGithub>=2.3.0
python-dotenv>=1.0.0

# GitHub webhook bot (`PYTHONPATH=src uvicorn github_bot.app:app`).
fastapi>=0.115.0
uvicorn[standard]>=0.32.0

# GitHub App JWT / installation tokens (`src/github_bot/auth.py`).
PyJWT[crypto]>=2.8.0

# Webhook idempotency (`src/github_bot/delivery_cache.py`); default ``redis://127.0.0.1:6379/0``.
redis>=5.0.0

# GitHub HTTP client with rate-limit handling (`src/github_bot/http_client.py`).
httpx>=0.27.0

# Learned style preferences (`.local/style_prefs.yaml` — `github_bot.style_prefs`).
PyYAML>=6.0.0

# OpenTelemetry tracing (`src/observability/tracer.py`) — OTLP HTTP to Phoenix / collector / LangSmith OTLP.
opentelemetry-api>=1.29.0
opentelemetry-sdk>=1.29.0
opentelemetry-exporter-otlp-proto-http>=1.29.0

# Local review vector memory (Chroma persistent DB + Ollama embeddings via VectorMemory).
chromadb>=0.5.0

# RepoGraph: tree-sitter import graph + JSON cache + LLM topology summary.
tree-sitter>=0.21.0
tree-sitter-python
tree-sitter-javascript
tree-sitter-typescript
